# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
  timestamp: string;
}

export interface User {
  id: string;
  firebaseUid: string;
  displayName: string;
  email: string;
  /** @nullable */
  photoUrl?: string | null;
  provider: string;
  createdAt: string;
}

export interface AuthSyncInput {
  displayName: string;
  email: string;
  /** @nullable */
  photoUrl?: string | null;
  provider: string;
}

export interface AiPathCoordinate {
  lat: number;
  lng: number;
  confidence: string;
  reasoning: string;
}

export interface AiPathCache {
  coordinates: AiPathCoordinate[];
  summary: string;
  generatedAt: string;
  approved: boolean;
}

export type PetStatus = (typeof PetStatus)[keyof typeof PetStatus];

export const PetStatus = {
  LOST: "LOST",
  FOUND: "FOUND",
  REUNITED: "REUNITED",
} as const;

export interface Pet {
  id: string;
  name: string;
  species: string;
  /** @nullable */
  breed?: string | null;
  color: string;
  size: string;
  /** @nullable */
  age?: string | null;
  description: string;
  photoUrl: string;
  status: PetStatus;
  lastSeenAddress: string;
  lastSeenLat: number;
  lastSeenLng: number;
  lastSeenAt: string;
  ownerId: string;
  aiPathCache?: AiPathCache | null;
  createdAt: string;
  updatedAt: string;
  owner?: User;
}

export type PetListItemStatus =
  (typeof PetListItemStatus)[keyof typeof PetListItemStatus];

export const PetListItemStatus = {
  LOST: "LOST",
  FOUND: "FOUND",
  REUNITED: "REUNITED",
} as const;

export interface PetListItem {
  id: string;
  name: string;
  species: string;
  /** @nullable */
  breed?: string | null;
  color: string;
  size: string;
  /** @nullable */
  age?: string | null;
  description?: string;
  photoUrl: string;
  status: PetListItemStatus;
  lastSeenAddress: string;
  lastSeenLat: number;
  lastSeenLng: number;
  lastSeenAt: string;
  ownerId: string;
  createdAt: string;
  updatedAt: string;
  owner?: User;
  sightingCount?: number;
}

export interface PetListResponse {
  pets: PetListItem[];
  total: number;
  page: number;
  limit: number;
}

export type SightingConfidence =
  (typeof SightingConfidence)[keyof typeof SightingConfidence];

export const SightingConfidence = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface Sighting {
  id: string;
  petId: string;
  reporterId: string;
  photoUrl: string;
  lat: number;
  lng: number;
  address: string;
  seenAt: string;
  /** @nullable */
  notes?: string | null;
  confidence: SightingConfidence;
  isFlagged: boolean;
  /** @nullable */
  flaggedAt?: string | null;
  sequence: number;
  createdAt: string;
  reporter?: User;
}

export type PetUpdateItemType =
  (typeof PetUpdateItemType)[keyof typeof PetUpdateItemType];

export const PetUpdateItemType = {
  GENERAL: "GENERAL",
  SIGHTING_FLAGGED: "SIGHTING_FLAGGED",
  SIGHTING_REMOVED: "SIGHTING_REMOVED",
  PET_FOUND: "PET_FOUND",
  PET_STILL_MISSING: "PET_STILL_MISSING",
  OWNER_NOTE: "OWNER_NOTE",
  AI_PATH_APPROVED: "AI_PATH_APPROVED",
  AI_PATH_DENIED: "AI_PATH_DENIED",
} as const;

export interface PetUpdateItem {
  id: string;
  petId: string;
  authorId: string;
  content: string;
  type: PetUpdateItemType;
  createdAt: string;
  author?: User;
}

export type PetDetailStatus =
  (typeof PetDetailStatus)[keyof typeof PetDetailStatus];

export const PetDetailStatus = {
  LOST: "LOST",
  FOUND: "FOUND",
  REUNITED: "REUNITED",
} as const;

export interface PetDetail {
  id: string;
  name: string;
  species: string;
  /** @nullable */
  breed?: string | null;
  color: string;
  size: string;
  /** @nullable */
  age?: string | null;
  description: string;
  photoUrl: string;
  status: PetDetailStatus;
  lastSeenAddress: string;
  lastSeenLat: number;
  lastSeenLng: number;
  lastSeenAt: string;
  ownerId: string;
  aiPathCache?: AiPathCache | null;
  createdAt: string;
  updatedAt: string;
  owner?: User;
  sightings: Sighting[];
  updates: PetUpdateItem[];
}

export type PetInputSpecies =
  (typeof PetInputSpecies)[keyof typeof PetInputSpecies];

export const PetInputSpecies = {
  dog: "dog",
  cat: "cat",
  other: "other",
} as const;

export type PetInputSize = (typeof PetInputSize)[keyof typeof PetInputSize];

export const PetInputSize = {
  small: "small",
  medium: "medium",
  large: "large",
} as const;

export interface PetInput {
  /** @minLength 1 */
  name: string;
  species: PetInputSpecies;
  /** @nullable */
  breed?: string | null;
  /** @minLength 1 */
  color: string;
  size: PetInputSize;
  /** @nullable */
  age?: string | null;
  /** @minLength 1 */
  description: string;
  photoUrl: string;
  /** @minLength 1 */
  lastSeenAddress: string;
  lastSeenLat: number;
  lastSeenLng: number;
  lastSeenAt: string;
}

export interface PetPatch {
  name?: string;
  species?: string;
  /** @nullable */
  breed?: string | null;
  color?: string;
  size?: string;
  /** @nullable */
  age?: string | null;
  description?: string;
  photoUrl?: string;
}

export type PetStatusInputStatus =
  (typeof PetStatusInputStatus)[keyof typeof PetStatusInputStatus];

export const PetStatusInputStatus = {
  LOST: "LOST",
  FOUND: "FOUND",
  REUNITED: "REUNITED",
} as const;

export interface PetStatusInput {
  status: PetStatusInputStatus;
}

export type SightingInputConfidence =
  (typeof SightingInputConfidence)[keyof typeof SightingInputConfidence];

export const SightingInputConfidence = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface SightingInput {
  photoUrl: string;
  lat: number;
  lng: number;
  /** @minLength 1 */
  address: string;
  seenAt: string;
  /** @nullable */
  notes?: string | null;
  confidence: SightingInputConfidence;
}

export interface PetUpdateInput {
  /** @minLength 1 */
  content: string;
}

export interface UploadResult {
  url: string;
}

export type ListPetsParams = {
  /**
   * @nullable
   */
  species?: string | null;
  /**
   * @nullable
   */
  status?: string | null;
  /**
   * @nullable
   */
  page?: number | null;
  /**
   * @nullable
   */
  limit?: number | null;
};