# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  title: Api
  version: "1.0.0"
servers:
  - url: /api
paths:
...
  /pets/{petId}/sightings:
    get:
      operationId: listSightings
      summary: List sightings for a pet
      tags: [sightings]
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of sightings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sighting"
    post:
      operationId: createSighting
      summary: Submit a new sighting
      tags: [sightings]
      security:
        - bearerAuth: []
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SightingInput"
      responses:
        "201":
          description: Sighting created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sighting"
        "401":
          description: Unauthorized
  /sightings/{id}/flag:
    patch:
      operationId: flagSighting
      summary: Flag sighting as invalid (owner only)
      tags: [sightings]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Flagged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sighting"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /sightings/{id}/unflag:
    patch:
      operationId: unflagSighting
      summary: Unflag a sighting (owner only)
      tags: [sightings]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unflagged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sighting"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /pets/{petId}/updates:
    get:
      operationId: listPetUpdates
      summary: Get pet timeline updates
      tags: [updates]
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of updates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PetUpdateItem"
    post:
      operationId: createPetUpdate
      summary: Post owner note
      tags: [updates]
      security:
        - bearerAuth: []
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PetUpdateInput"
      responses:
        "201":
          description: Update posted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PetUpdateItem"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
  schemas:
...
    User:
      type: object
      required: [id, firebaseUid, displayName, email, provider, createdAt]
      properties:
        id:
          type: string
        firebaseUid:
          type: string
        displayName:
          type: string
        email:
          type: string
        photoUrl:
          type: ["string", "null"]
        provider:
          type: string
        createdAt:
          type: string
...
    Sighting:
      type: object
      required: [id, petId, reporterId, photoUrl, lat, lng, address, seenAt, confidence, isFlagged, sequence, createdAt]
      properties:
        id:
          type: string
        petId:
          type: string
        reporterId:
          type: string
        photoUrl:
          type: string
        lat:
          type: number
        lng:
          type: number
        address:
          type: string
        seenAt:
          type: string
        notes:
          type: ["string", "null"]
        confidence:
          type: string
          enum: [high, medium, low]
        isFlagged:
          type: boolean
        flaggedAt:
          type: ["string", "null"]
        sequence:
          type: integer
        createdAt:
          type: string
        reporter:
          $ref: "#/components/schemas/User"
    PetUpdateItem:
      type: object
      required: [id, petId, authorId, content, type, createdAt]
      properties:
        id:
          type: string
        petId:
          type: string
        authorId:
          type: string
        content:
          type: string
        type:
          type: string
          enum: [GENERAL, SIGHTING_FLAGGED, SIGHTING_REMOVED, PET_FOUND, PET_STILL_MISSING, OWNER_NOTE, AI_PATH_APPROVED, AI_PATH_DENIED]
        createdAt:
          type: string
        author:
          $ref: "#/components/schemas/User"
...
    SightingInput:
      type: object
      required: [photoUrl, lat, lng, address, seenAt, confidence]
      properties:
        photoUrl:
          type: string
        lat:
          type: number
        lng:
          type: number
        address:
          type: string
          minLength: 1
        seenAt:
          type: string
        notes:
          type: ["string", "null"]
        confidence:
          type: string
          enum: [high, medium, low]
    PetUpdateInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
...