# lib/api-zod/src/generated/types/sightingInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
import type { SightingInputConfidence } from "./sightingInputConfidence";

export interface SightingInput {
  photoUrl: string;
  lat: number;
  lng: number;
  /** @minLength 1 */
  address: string;
  seenAt: string;
  /** @nullable */
  notes?: string | null;
  confidence: SightingInputConfidence;
}